wip: play around a bit with 1786, got some ideas from dp solutions
Somehow it performs worse in Leetcode test suite when it comes to runtime, but consumes a lot less memory than the initial solution.
Performance is better than the initial one was in test cases that I have.
Perhaps Leetcode has a test where graph is very large and partitioned by last node that comes somewhere very late in the suite.