wip: updated 1786 for better Leetcode performance
Apparently, the order in which Leetcode sends you edges matters, and it is much more efficient to iterate nodes in that specific order.
I have attempted to sort edges by weight, but that was a mistake.
Now the code performs the same as other DP solution and is more or less the same.
Notably, I found that providing a Comparable record to PriorityQueue is a tiny bit faster than relying on Comparator, which is weird.